PREPROCESSING/HE_DECONV/requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
ctk-cli==1.5
cycler==0.10.0
dask==2.12.0
decorator==4.4.2
diskcache==4.1.0
distributed==2.12.0
fsspec==0.6.2
gast==0.2.2
girder-client==3.0.10
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
HeapDict==1.0.1
histomicstk==1.0.3.dev63
idna==2.9
imageio==2.8.0
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
large-image==1.0.4
large-image-source-ometiff==1.0.4
large-image-source-openslide==1.0.4
large-image-source-pil==1.0.4
large-image-source-tiff==1.0.4
libtiff==0.4.2
Markdown==3.2.1
matplotlib==3.2.0
msgpack==1.0.0
networkx==2.4
nimfa==1.4.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
openslide-python==1.1.1
opt-einsum==3.2.0
pandas==1.0.1
Pillow==7.0.0
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
sortedcontainers==2.1.0
tblib==1.6.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.4
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
zict==2.0.0
